Publicize: collapse the legacy/dashboard connection component duality
Every connection surface shipped twice: a `@wordpress/components` variant
for the block editor sidebar and a `@wordpress/ui` (WPDS) `-dashboard`
variant for the Social page, picked at runtime by `useIsDashboard()`. Keep
the WPDS implementation as the single one and delete the legacy twin for
the modal, the services list/item/inputs/connection-info, and the
connection-management row.

`useIsDashboard()` survives in `connection-template`, where it selects
behaviour rather than a component variant: it gates the dashboard-only
upsell and the global-template read.

Stylesheets are deliberately left split in this commit — the kept
components still import `style-dashboard.module.scss`. Collapsing them
carries the only visual risk in this change, so it lands separately.

Specs are merged rather than dropped: the connection-management page
objects move onto the WPDS affordances (a single `Collapsible` toggle and
`data-panel-open` in place of "Open panel"/"Close panel" and `.is-opened`),
and the legacy service-connection-info spec's mark-as-shared case is
carried into the merged file.
