+++ projects/packages/publicize/_inc/components/connection-management/index.tsx
@@ -3,15 +3,12 @@ import { useDispatch, useSelect } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
 import { Button } from '@wordpress/ui';
 import clsx from 'clsx';
-import { useIsDashboard } from '../../hooks/use-is-dashboard';
 import { useUserCanShareConnection } from '../../hooks/use-user-can-share-connection';
 import { store } from '../../social-store';
 import { ThemedConnectionsModal as ManageConnectionsModal } from '../manage-connections-modal';
 import { useService } from '../services/use-service';
 import { ConnectionInfo } from './connection-info';
-import { DashboardConnectionInfo } from './connection-info-dashboard';
-import dashboardStyles from './style-dashboard.module.scss';
-import styles from './style.module.scss';
+import listStyles from './style-dashboard.module.scss';
 
 const ConnectionManagement = ( {
 	className = null,
@@ -19,13 +16,6 @@ const ConnectionManagement = ( {
 	hideConnectButton = false,
 	hideHeading = false,
 } ) => {
-	const isDashboard = useIsDashboard();
-	const ConnectionInfoVariant = isDashboard ? DashboardConnectionInfo : ConnectionInfo;
-	// The Social dashboard owns its list chrome (edge-to-edge dividers, no
-	// outline, rows supply their own padding). The block editor keeps the trunk
-	// `style.module.scss` classes byte-for-byte.
-	const listStyles = isDashboard ? dashboardStyles : styles;
-
 	const {
 		connections: rawConnections,
 		deletingConnections,
@@ -82,7 +72,7 @@ const ConnectionManagement = ( {
 									key={ connection.connection_id }
 								>
 									<Disabled isDisabled={ isUpdatingOrDeleting }>
-										<ConnectionInfoVariant
+										<ConnectionInfo
 											connection={ connection }
 											service={ getService( connection.service_name ) }
 											canMarkAsShared={ canMarkAsShared }
