+++ projects/packages/publicize/_inc/components/connection-management/tests/pageObjects/ConnectionManagementPage.js
@@ -21,8 +21,16 @@ export class ConnectionManagementPageObject {
 		return this.container.queryAllByRole( 'presentation', { name: 'Loading account details' } );
 	}
 
+	/*
+	 * The WPDS row renders the account name and the network label as separate
+	 * text nodes, and the fixtures reuse one string for both ("Facebook" on
+	 * Facebook), so a bare text query matches twice. The disclosure toggle's
+	 * accessible name is the row's unique, user-facing identity.
+	 */
 	getConnectionByName( name ) {
-		return this.container.queryByText( name );
+		return this.container.queryByRole( 'button', {
+			name: new RegExp( `^Toggle details for ${ name } on ` ),
+		} );
 	}
 
 	get connectionPanels() {
