+++ projects/packages/publicize/_inc/components/services/service-item-details.tsx
@@ -2,12 +2,10 @@ import { Disabled } from '@wordpress/components';
 import { useViewportMatch } from '@wordpress/compose';
 import { useSelect } from '@wordpress/data';
 import clsx from 'clsx';
-import { useIsDashboard } from '../../hooks/use-is-dashboard';
 import { useUserCanShareConnection } from '../../hooks/use-user-can-share-connection';
 import { store as socialStore } from '../../social-store';
 import { Connection } from '../../social-store/types';
 import { ServiceConnectionInfo } from './service-connection-info';
-import { DashboardServiceConnectionInfo } from './service-connection-info-dashboard';
 import styles from './style.module.scss';
 import { SupportedService } from './types';
 
@@ -24,10 +22,6 @@ export type ServicesItemDetailsProps = {
  * @return {import('react').ReactNode} Service item details component
  */
 export function ServiceItemDetails( { service, serviceConnections }: ServicesItemDetailsProps ) {
-	const isDashboard = useIsDashboard();
-	const ServiceConnectionInfoVariant = isDashboard
-		? DashboardServiceConnectionInfo
-		: ServiceConnectionInfo;
 	const isSmall = useViewportMatch( 'small', '<' );
 
 	const { deletingConnections, updatingConnections } = useSelect( select => {
@@ -52,7 +46,7 @@ export function ServiceItemDetails( { service, serviceConnections }: ServicesIte
 					return (
 						<li key={ connection.connection_id }>
 							<Disabled isDisabled={ isUpdatingOrDeleting }>
-								<ServiceConnectionInfoVariant
+								<ServiceConnectionInfo
 									connection={ connection }
 									service={ service }
 									canMarkAsShared={ canMarkAsShared }
