+++ projects/packages/publicize/_inc/components/services/services-list.tsx
@@ -1,9 +1,10 @@
 import { useSelect } from '@wordpress/data';
 import { useMemo } from '@wordpress/element';
+import { Card } from '@wordpress/ui';
 import { store } from '../../social-store';
 import { Connection } from '../../social-store/types';
 import { ServiceItem } from './service-item';
-import styles from './style.module.scss';
+import styles from './style-dashboard.module.scss';
 import { useSupportedServices } from './use-supported-services';
 
 /**
@@ -36,16 +37,18 @@ export function ServicesList() {
 	const reconnectingAccount = useSelect( select => select( store ).getReconnectingAccount(), [] );
 
 	return (
-		<ul className={ styles.services }>
-			{ supportedServices.map( service => (
-				<li key={ service.id } className={ styles[ 'service-list-item' ] }>
-					<ServiceItem
-						service={ service }
-						serviceConnections={ connections[ service.id ] || [] }
-						isPanelDefaultOpen={ reconnectingAccount?.service_name === service.id }
-					/>
-				</li>
-			) ) }
-		</ul>
+		<Card.Root>
+			<ul className={ styles.services }>
+				{ supportedServices.map( service => (
+					<li key={ service.id } className={ styles[ 'service-list-item' ] }>
+						<ServiceItem
+							service={ service }
+							serviceConnections={ connections[ service.id ] || [] }
+							isPanelDefaultOpen={ reconnectingAccount?.service_name === service.id }
+						/>
+					</li>
+				) ) }
+			</ul>
+		</Card.Root>
 	);
 }
