+++ projects/packages/publicize/_inc/components/services/tests/services-list.test.jsx
@@ -2,11 +2,11 @@ import { render, renderHook, screen } from '@testing-library/react';
 import { useSelect } from '@wordpress/data';
 import { store } from '../../../social-store';
 import { setup } from '../../../utils/test-factory';
-import { DashboardServicesList } from '../services-list-dashboard';
+import { ServicesList } from '../services-list';
 
-// Keep the list shallow — each row is covered by service-item-dashboard's own test.
-jest.mock( '../service-item-dashboard', () => ( {
-	DashboardServiceItem: ( { service } ) => <div>Row: { service.label }</div>,
+// Keep the list shallow — each row is covered by service-item's own test.
+jest.mock( '../service-item', () => ( {
+	ServiceItem: ( { service } ) => <div>Row: { service.label }</div>,
 } ) );
 
 const prepareStore = () => {
@@ -16,14 +16,14 @@ const prepareStore = () => {
 	jest.spyOn( storeSelect, 'getReconnectingAccount' ).mockReturnValue( undefined );
 };
 
-describe( 'DashboardServicesList', () => {
+describe( 'ServicesList', () => {
 	afterEach( () => {
 		jest.clearAllMocks();
 	} );
 
 	test( 'renders a row for each supported service', () => {
 		prepareStore();
-		render( <DashboardServicesList /> );
+		render( <ServicesList /> );
 
 		// The shared factory seeds Facebook, LinkedIn and Instagram Business.
 		expect( screen.getByText( 'Row: Facebook' ) ).toBeInTheDocument();
